Add support for other online XYZ sources (#9)
* Add OSM scaffolding

* Add support for Esri World Imagery and Google Hybrid XYZ

* Consolidate tile requests from different services
