+++ README.md
@@ -22,7 +22,7 @@ Options if `style` is "yes":
 *  `-i` or `--stylesources`: Directory where any local source files (GeoJSON, XYZ directory, MBTiles) specified in your provided style are located (required)
 
 Required options if `style` is "no":
-*  `-O` or `--onlinesource`: Specify an online source to be used as a background map (currently supported: "bing") (required)
+*  `-O` or `--onlinesource`: Specify an online source to be used as a background map (currently supported: "bing", "esri", "google") (required)
 *  `-a` or `--overlay`: Provide an GeoJSON object for a feature layer to overlay on top of the online source (required)
 *  `-k` or `--apikey`: API key that may be required for your online source
   
@@ -64,4 +64,6 @@ Three easy ways to examine and inspect the mbtiles:
 
 This tool makes it possible to download tiles from various API sources for offline usage. Here are links to the licensing and API limitations for each source:
 
-1. Bing: [Terms of Use](https://www.microsoft.com/en-us/maps/bing-maps/product) and information on [accessing Bing Maps tiles](https://learn.microsoft.com/en-us/bingmaps/rest-services/directly-accessing-the-bing-maps-tiles)
+1. Bing Satellite: API [Terms of Use](https://www.microsoft.com/en-us/maps/bing-maps/product) and information on [accessing Bing Maps tiles](https://learn.microsoft.com/en-us/bingmaps/rest-services/directly-accessing-the-bing-maps-tiles)
+2. Esri World Imagery (for Export): [Terms of use](https://www.arcgis.com/home/item.html?id=226d23f076da478bba4589e7eae95952)
+3. Google Hybrid: API [Terms of Use](https://developers.google.com/maps/documentation/tile/policies)
