+++ package.json
@@ -39,5 +39,5 @@
   },
   "keywords": [],
   "author": "Rudo Kemper",
-  "license": "GPL-3.0"
+  "license": "MIT"
 }
